More modularization of Functions codebase (#259)
